+++ src/stores/CounterStore.tsx
@@ -1,0 +1,11 @@
+import { createStore } from "solid-js/store";
+
+export type Counter = {
+    rendered: number,
+    validated: number
+}
+
+export const [counter, setCounter] = createStore<Counter>({
+    rendered: 0,
+    validated: 0
+});